fix check_if_token_is_service_account
