+++ litellm/proxy/proxy_config.yaml
@@ -4,12 +4,6 @@ model_list:
       model: openai/gpt-4o
       api_key: sk-xxxxxxx
 
-mcp_servers:
-  {
-    "zapier_mcp": {
-      "url": "https://actions.zapier.com/mcp/sk-akxxxxx/sse"
-    },
-    "fetch": {
-      "url": "http://localhost:8000/sse"
-    }
-  }
+general_settings:
+    service_account_settings: 
+        enforced_params: ["user"] # this means the "user" param is enforced for all requests made through any service account keys
